A template exercice
We had to, for a day, reproduce the perfect example of a CV site that had been given to us that morning. Here is my creation. Hello and good luck to Philippe who will pass by here. Some information in this site is not up to date, this just serves as an example. All uses are possible, so treat yourself !